php project

+++ project/admin.php
@@ -1,14 +1,16 @@
 <?php
-include("vendor/autoload.php");
 
+use Helpers\Auth;
 use Libs\Database\MySQL;
 use Libs\Database\UsersTable;
-use Helpers\Auth;
 
-$auth = Auth::check();
+include("vendor/autoload.php");
 
 $table = new UsersTable(new MySQL);
 $users = $table->all();
+
+$auth = Auth::check();
+
 ?>
 <!DOCTYPE html>
 <html lang="en">
@@ -18,6 +20,7 @@ $users = $table->all();
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Admin</title>
     <link rel="stylesheet" href="css/bootstrap.min.css">
+    <script src="js/bootstrap.bundle.min.js" defer></script>
 </head>
 
 <body>
@@ -67,8 +70,33 @@ $users = $table->all();
                         <?php endif ?>
                     </td>
                     <td>
-                        <div class="btn-group">
-                            <a href="_actions/delete.php?id=<?= $user->id ?>" class="btn btn-sm btn-outline-danger">Delete</a>
+                        <div class="btn-group dropdown">
+                            <?php if ($auth->role_id == 3): ?>
+                                <a href="#" class="btn btn-sm btn-outline-primary dropdown-toggle"
+                                    data-bs-toggle="dropdown">Role</a>
+                                <div class="dropdown-menu">
+                                    <a href="_actions/role.php?id=<?= $user->id ?>&role=1"
+                                        class="dropdown-item">User</a>
+                                    <a href="_actions/role.php?id=<?= $user->id ?>&role=2"
+                                        class="dropdown-item">Manager</a>
+                                    <a href="_actions/role.php?id=<?= $user->id ?>&role=3"
+                                        class="dropdown-item">Admin</a>
+                                </div>
+                            <?php endif ?>
+
+                            <?php if ($auth->role_id >= 2): ?>
+                                <?php if ($user->suspended): ?>
+                                    <a href="_actions/unsuspend.php?id=<?= $user->id ?>"
+                                        class="btn btn-sm btn-warning">Ban</a>
+                                <?php else: ?>
+                                    <a href="_actions/suspend.php?id=<?= $user->id ?>"
+                                        class="btn btn-sm btn-outline-warning">Ban</a>
+                                <?php endif ?>
+                            <?php endif ?>
+
+                            <?php if ($auth->role_id == 3): ?>
+                                <a href="_actions/delete.php?id=<?= $user->id ?>" class="btn btn-sm btn-outline-danger">Delete</a>
+                            <?php endif ?>
                         </div>
                     </td>
                 </tr>
